+++ server/routes/index.js
@@ -1,0 +1,19 @@
+const fs = require('fs')
+const path = require('path')
+
+const basename = path.basename(module.filename)
+const logger = require('../etc/logger')
+
+const getRoutes = app => {
+  fs
+    .readdirSync(__dirname)
+    .filter(file => file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js')
+    .map(file => file.replace('.js', ''))
+    .forEach(file => {
+      const route = require(`./${file}`)
+      app.use(`/api/${file}`, route)
+      logger.info(`registering route /api/${file}`)
+    })
+}
+
+module.exports = getRoutes
